requirements.txt
langchain==0.1.11
langchain-core==0.1.52
langchain_openai==0.1.7
python-docx==1.1.2
python-dotenv==1.0.1
openai==1.35.10
instructor==1.3.4
langchain-chroma==0.1.2
sentence-transformers==3.0.1
streamlit==1.36.0
uvicorn==0.30.1
fastapi==0.111.0
openpyxl==3.1.5